requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
PyPDF2==3.0.1
pydantic==2.3.0
python-dotenv==1.0.0
langchain-groq==0.0.1
langchain-core==0.1.1
langchain-community==0.0.9